backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1